+++ packet_channel_open.c
@@ -36,7 +36,7 @@ int packet_channel_open(struct buf *b1, struct buf *b2) {
                             &remotewindow); /* uint32    initial window size */
     pos = packetparser_uint32(b1->buf, b1->len, pos,
                               &maxpacket); /* uint32    maximum packet size */
-    if (maxpacket > PACKET_LIMIT) maxpacket = PACKET_LIMIT;
+    if (maxpacket > CHANNEL_PACKET_LIMIT) maxpacket = CHANNEL_PACKET_LIMIT;
     if (maxpacket < 32) maxpacket = 32;
 
     reason = SSH_OPEN_UNKNOWN_CHANNEL_TYPE;
@@ -68,13 +68,8 @@ int packet_channel_open(struct buf *b1, struct buf *b2) {
             buf_putnum32(b2, id);          /* uint32    recipient channel */
             buf_putnum32(b2, id);          /* uint32    sender channel */
             buf_putnum32(b2, localwindow); /* uint32    initial window size */
-            /*
-            XXX
-            use PACKET_LIMIT/2 as maximum packet size,
-            workaround for miscalculated packet_length
-            */
-            buf_putnum32(b2,
-                         PACKET_LIMIT / 2); /* uint32    maximum packet size */
+            buf_putnum32(
+                b2, CHANNEL_PACKET_LIMIT); /* uint32    maximum packet size */
             packet_put(b2);
             buf_purge(b2);
             return 1;