+++ packet_channel_send.c
@@ -15,17 +15,19 @@ Public domain.
 void packet_channel_send_data(struct buf *b) {
 
     long long r;
-    long long limit = PACKET_LIMIT;
+    long long limit = CHANNEL_PACKET_LIMIT;
 
     if (channel.maxpacket > 0) {
-        if (channel.maxpacket < PACKET_LIMIT) { limit = channel.maxpacket; }
+        if (channel.maxpacket < CHANNEL_PACKET_LIMIT) {
+            limit = channel.maxpacket;
+        }
     }
 
     buf_purge(b);
 
-    if (b->alloc <= PACKET_LIMIT) bug_nomem();
+    if (b->alloc < CHANNEL_PACKET_LIMIT + 9) bug_nomem();
     if (!packet_putisready()) return;
-    r = channel_read(b->buf + 9, limit - 9);
+    r = channel_read(b->buf + 9, limit);
     if (r == 0) return;
     b->len = r + 9;
     b->buf[0] = SSH_MSG_CHANNEL_DATA; /* byte      SSH_MSG_CHANNEL_DATA */
@@ -40,17 +42,19 @@ void packet_channel_send_data(struct buf *b) {
 void packet_channel_send_extendeddata(struct buf *b) {
 
     long long r;
-    long long limit = PACKET_LIMIT;
+    long long limit = CHANNEL_PACKET_LIMIT;
 
     if (channel.maxpacket > 0) {
-        if (channel.maxpacket < PACKET_LIMIT) { limit = channel.maxpacket; }
+        if (channel.maxpacket < CHANNEL_PACKET_LIMIT) {
+            limit = channel.maxpacket;
+        }
     }
 
     buf_purge(b);
 
-    if (b->alloc <= PACKET_LIMIT) bug_nomem();
+    if (b->alloc < CHANNEL_PACKET_LIMIT + 13) bug_nomem();
     if (!packet_putisready()) return;
-    r = channel_extendedread(b->buf + 13, limit - 13);
+    r = channel_extendedread(b->buf + 13, limit);
     if (r == 0) return;
     b->len = r + 13;
     b->buf[0] =